# packages/api-client/openapi/finora.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "operationId": "categoriesList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Номер страницы от 1",
            "schema": {
              "minimum": 1,
              "maximum": 9999999,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Размер страницы",
            "schema": {
              "default": 25,
              "type": "number",
              "enum": [
                10,
                25,
                50
              ]
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Состояние категорий",
            "schema": {
              "default": "all",
              "type": "string",
              "enum": [
                "all",
                "active",
                "archived"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Список собственных записей с серверной пагинацией",
        "tags": [
          "Категории"
        ]
      },
      "post": {
        "operationId": "categoriesCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Создать запись с атомарным аудитом",
        "tags": [
          "Категории"
        ]
      }
    },
    "/api/v1/categories/options": {
      "get": {
        "operationId": "categoriesOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Собственный компактный справочник для выбора и фильтрации; включает архив для истории",
        "tags": [
          "Категории"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "categoriesGet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получить собственную запись",
        "tags": [
          "Категории"
        ]
      },
      "patch": {
        "operationId": "categoriesUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryPatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Изменить собственную запись; требуется хотя бы одно разрешённое поле",
        "tags": [
          "Категории"
        ]
      },
      "delete": {
        "operationId": "categoriesDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRemovalDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса; 429 содержит Retry-After в секундах",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Удалить собственную запись с сохранением аудита",
        "tags": [
          "Категории"
        ]
      }
    },
...
  },
  "info": {
    "title": "Finora API",
    "description": "Finora: авторизация, настройки, категории и операции. Изменяющие запросы требуют разрешённый Origin. Сессия: HttpOnly cookie, 24 часа.",
    "version": "0.5.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "finora_session"
      }
    },
    "schemas": {
      "ProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Категория ошибки",
            "example": "internal_error"
          },
          "title": {
            "type": "string",
            "description": "Краткое описание ошибки"
          },
          "status": {
            "type": "number",
            "description": "HTTP-статус"
          },
          "detail": {
            "type": "string",
            "description": "Безопасное описание причины"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Ошибки полей"
          },
          "traceId": {
            "type": "string",
            "description": "Идентификатор запроса"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "errors",
          "traceId"
        ]
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название категории; trim, уникально без регистра в пределах владельца и типа",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "description": "Тип категории",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "icon": {
            "type": "string",
            "description": "Иконка категории",
            "enum": [
              "briefcase-business",
              "laptop",
              "shopping-basket",
              "house",
              "tram-front",
              "utensils",
              "train-front",
              "repeat",
              "heart",
              "gift",
              "graduation-cap",
              "wallet"
            ]
          },
          "color": {
            "type": "string",
            "description": "Цвет категории",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid"
          },
          "archivedAt": {
            "type": "string",
            "description": "Дата архивирования",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Создана",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Изменена",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "type",
          "icon",
          "color",
          "id",
          "archivedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CategoryPageDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "Номер страницы"
          },
          "pageSize": {
            "type": "number",
            "description": "Размер страницы"
          },
          "total": {
            "type": "number",
            "description": "Всего собственных записей после фильтрации"
          },
          "items": {
            "description": "Категории страницы",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            }
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "items"
        ]
      },
      "CategoryInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название категории; trim, уникально без регистра в пределах владельца и типа",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "description": "Тип категории",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "icon": {
            "type": "string",
            "description": "Иконка категории",
            "enum": [
              "briefcase-business",
              "laptop",
              "shopping-basket",
              "house",
              "tram-front",
              "utensils",
              "train-front",
              "repeat",
              "heart",
              "gift",
              "graduation-cap",
              "wallet"
            ]
          },
          "color": {
            "type": "string",
            "description": "Цвет категории",
            "pattern": "^#[0-9a-fA-F]{6}$"
          }
        },
        "required": [
          "name",
          "type",
          "icon",
          "color"
        ],
        "additionalProperties": false
      },
      "CategoryPatchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название категории; trim, уникально без регистра в пределах владельца и типа",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "description": "Тип категории",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "icon": {
            "type": "string",
            "description": "Иконка категории",
            "enum": [
              "briefcase-business",
              "laptop",
              "shopping-basket",
              "house",
              "tram-front",
              "utensils",
              "train-front",
              "repeat",
              "heart",
              "gift",
              "graduation-cap",
              "wallet"
            ]
          },
          "color": {
            "type": "string",
            "description": "Цвет категории",
            "pattern": "^#[0-9a-fA-F]{6}$"
          }
        },
        "additionalProperties": false
      },
      "CategoryRemovalDto": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "description": "Результат: удалена неиспользованная либо архивирована используемая категория; активные связанные правила также архивируются",
            "enum": [
              "deleted",
              "archived"
            ]
          }
        },
        "required": [
          "outcome"
        ]
      },
...
    }
  }
}